Enable bundle tests for React.Fragment
